+++ app/src/js/components/Pdr/list.js
@@ -12,12 +12,8 @@ import {
 } from '../../actions';
 import { lastUpdated, tally, displayCase } from '../../utils/format';
 import {
-  tableHeader,
-  tableRow,
-  tableSortProps,
-  errorTableHeader,
-  errorTableRow,
-  errorTableSortProps,
+  tableColumns,
+  errorTableColumns,
   bulkActions
 } from '../../utils/table-config/pdrs';
 import Dropdown from '../DropDown/dropdown';
@@ -87,12 +83,10 @@ class ActivePdrs extends React.Component {
             list={list}
             dispatch={this.props.dispatch}
             action={listPdrs}
-            tableHeader={view === 'failed' ? errorTableHeader : tableHeader}
-            tableRow={view === 'failed' ? errorTableRow : tableRow}
-            tableSortProps={view === 'failed' ? errorTableSortProps : tableSortProps}
+            tableColumns={view === 'failed' ? errorTableColumns : tableColumns}
             query={this.generateQuery()}
             bulkActions={this.generateBulkActions()}
-            rowId={'pdrName'}
+            rowId='pdrName'
           />
         </section>
       </div>