+++ app/src/js/utils/table-config/services.js
@@ -1,18 +1,28 @@
 'use strict';
 import { tally } from '../format';
 
-export const tableHeader = [
-  'Service Name',
-  'Status',
-  'Desired Tasks',
-  'Pending Tasks',
-  'Running Tasks'
-];
-
-export const tableRow = [
-  'name',
-  'status',
-  (d) => tally(d['desiredCount']),
-  (d) => tally(d['pendingCount']),
-  (d) => tally(d['runningCount'])
+export const tableColumns = [
+  {
+    Header: 'Service Name',
+    accessor: 'name'
+  },
+  {
+    Header: 'Status',
+    accessor: 'status'
+  },
+  {
+    Header: 'Desired Tasks',
+    accessor: row => tally(row['desiredCount']),
+    id: 'desiredCount'
+  },
+  {
+    Header: 'Pending Tasks',
+    accessor: row => tally(row['pendingCount']),
+    id: 'pendingCount'
+  },
+  {
+    Header: 'Running Tasks',
+    accessor: row => tally(row['runningCount']),
+    id: 'runningCount'
+  }
 ];